feat(components) Introduce DCC Descriptor
