+++ src/app/app.component.html
@@ -328,18 +328,12 @@
   justify-content: space-around; 
   align-items: stretch
 }
+
 .dcc-content-qr > *{
   flex-grow: 1;
   flex-shrink: 1;
   margin: 1em;
   width: 100%;
-}
-
-.dcc-content-qr > *,
-.dcc-content-submit > *{
-  flex-grow: 1;
-  flex-shrink: 1;
-  margin: 1em;
 }
 
 .dcc-footer {
@@ -376,31 +370,10 @@
     <!-- Content -->
     <div class="dcc-content">
       <div class="dcc-content-qr">
+        <!-- DCC QR Code for scanning -->
         <app-dccqr></app-dccqr>
-
-
-        <mat-card>
-          <mat-card-content>
-            <mat-tree class="dcc-qr-description"[dataSource]="dataSource2" [treeControl]="treeControl" style="flex-grow: 1; flex-shrink: 1">
-              <!-- This is the tree node template for leaf nodes -->
-              <mat-tree-node *matTreeNodeDef="let node" matTreeNodePadding>
-                <!-- use a disabled button to provide padding for tree leaf -->
-                <button mat-icon-button disabled></button>
-                {{node.name}}
-              </mat-tree-node>
-              <!-- This is the tree node template for expandable nodes -->
-              <mat-tree-node *matTreeNodeDef="let node;when: hasChild" matTreeNodePadding>
-                <button mat-icon-button matTreeNodeToggle
-                        [attr.aria-label]="'Toggle ' + node.name">
-                  <mat-icon class="mat-icon-rtl-mirror">
-                    {{treeControl.isExpanded(node) ? 'expand_more' : 'chevron_right'}}
-                  </mat-icon>
-                </button>
-                {{node.name}}
-              </mat-tree-node>
-            </mat-tree>
-          </mat-card-content>
-        </mat-card>
+        <!-- DCC Details -->
+        <app-dccdescription></app-dccdescription>
       </div>
 
       <div class="dcc-content-submit">
